Big overhaul of 64-bit kernel along the lines of what we already have for the 64-bit kernel just more radical.

+++ include/asm-mips/mipsregs.h
@@ -466,7 +466,7 @@
  * For now use this only with interrupts disabled!
  */
 #define read_64bit_cp0_register(source)                         \
-({ int __res;                                                   \
+({ unsigned long __res;                                         \
         __asm__ __volatile__(                                   \
         ".set\tmips3\n\t"                                       \
         "dmfc0\t%0,"STR(source)"\n\t"                           \
